requirements.txt
Flask
pymongo==4.10.1
python-dotenv
requests
faiss-cpu
numpy
sentence-transformers
fpdf
huggingface-hub
gunicorn